refactor: split into an importable library and a cmd/simslim CLI
The repo root becomes `package simslim`, an importable library holding every
piece of slimming logic; the CLI moves to `cmd/simslim`. Previously everything
was `package main`, so nothing here could be imported at all — `package main`
is not importable regardless of what it exports.

Consumers now get `import "github.com/mobai-app/simslim"` with no stutter, and
the library carries no external dependencies: urfave/cli stays on the CLI side.

Terminal-facing code moved to cmd/simslim: the interactive profile wizard
(split out of profile_file.go into wizard.go), writeJSON, truncate/humanBytes,
fatal and usage. The library never writes to stdout — it returns values and
reports progress through the Reporter callback the caller supplies.

Two name collisions with existing types forced renames: status() is now
ReadStatus (and statusForDevice ReadStatusForDevice), and diskCleanupPlan() is
now PlanDiskCleanup. The shutdown/boot timeouts move to simctl.go as exported
ShutdownTimeout/BootTimeout, since both packages need them.

`go install` gains a path element: github.com/mobai-app/simslim/cmd/simslim.
The installed binary is still named simslim, and -X main.version still applies.

CLI behavior is unchanged: JSON output, human output, help text and exit codes
are byte-identical to the previous build, and the test inventory is unchanged.

extraDeviceSets remains a package-level global for now, with ExtraDeviceSetTokens
and ResetDeviceSets added so the CLI's flag tests can still drive it. Making
device sets caller-owned state is a follow-up.
