+++ disk_inventory.go
@@ -1,4 +1,4 @@
-package main
+package simslim
 
 import (
 	"fmt"
