+++ disk_inventory_test.go
@@ -1,4 +1,4 @@
-package main
+package simslim
 
 import (
 	"path/filepath"
